Fix function line limit
