+++ Rocket.Chat/Controllers/Subscriptions/NewRoomViewController.swift
@@ -189,7 +189,7 @@ class NewRoomViewController: BaseViewController {
                 }
             }
         }, errored: { _ in
-            // TODO: Handle error
+            Alert.defaultError.present()
         })
     }
